requirements.txt
torch
torchvision
numpy
matplotlib
Pillow
scikit-image
scikit-learn
tqdm
efficientnet_pytorch